Review : merge 2 device & 2 sensor classes
