+++ homeassistant/components/freebox/device_tracker.py
@@ -1,4 +1,5 @@
 """Support for Freebox devices (Freebox v6 and Freebox mini 4K)."""
+from datetime import datetime
 import logging
 from typing import Dict
 
@@ -9,8 +10,7 @@ from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.typing import HomeAssistantType
 
-from .const import DOMAIN, TRACKER_UPDATE
-from .router import FreeboxDevice
+from .const import DEFAULT_DEVICE_NAME, DOMAIN, TRACKER_UPDATE
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -24,66 +24,113 @@ async def async_setup_entry(
     hass: HomeAssistantType, entry: ConfigEntry, async_add_entities
 ) -> None:
     """Set up the device_tracker."""
-    for device in hass.data[DOMAIN].devices.values():
-        _LOGGER.debug("Adding device_tracker for %s", device.name)
-        async_add_entities([FreeboxTrackerEntity(device)])
+    fbx = hass.data[DOMAIN]
+
+    entities = []
+
+    for device in fbx.devices.values():
+        entities.append(device)
+
+    async_add_entities(entities)
 
 
-class FreeboxTrackerEntity(TrackerEntity):
-    """Represent a tracked device."""
+class FreeboxDevice(TrackerEntity):
+    """Representation of a Freebox device."""
 
-    def __init__(self, device: FreeboxDevice):
-        """Set up the Freebox tracker entity."""
-        self._device = device
+    def __init__(self, device: Dict[str, any]):
+        """Initialize a Freebox device."""
+        self._name = device["primary_name"].strip() or DEFAULT_DEVICE_NAME
+        self._mac = device["l2ident"]["id"]
+        self._manufacturer = device["vendor_name"]
+        self._icon = icon_for_freebox_device(device)
         self._unsub_dispatcher = None
+
+        self.update_state(device)
+
+    def update_state(self, device: Dict[str, any]) -> None:
+        """Update the Freebox device."""
+        self._active = device["active"]
+        if device.get("attrs") is None:
+            # device
+            self._reachable = device["reachable"]
+            self._attrs = {
+                "reachable": self._reachable,
+                "last_time_reachable": datetime.fromtimestamp(
+                    device["last_time_reachable"]
+                ),
+                "last_time_activity": datetime.fromtimestamp(device["last_activity"]),
+            }
+        else:
+            # router
+            self._attrs = device["attrs"]
 
     @property
     def unique_id(self) -> str:
         """Return a unique ID."""
-        return self._device.mac
+        return self.mac
 
     @property
     def name(self) -> str:
-        """Return the name of the device."""
-        return self._device.name
+        """Return the name."""
+        return self._name
 
     @property
     def latitude(self):
-        """Return latitude value of the device."""
-        if self._device.active:
+        """Return the latitude."""
+        if self.active:
             return self.hass.config.latitude
         return None
 
     @property
     def longitude(self):
-        """Return longitude value of the device."""
-        if self._device.active:
+        """Return the longitude."""
+        if self.active:
             return self.hass.config.longitude
         return None
 
     @property
     def source_type(self) -> str:
-        """Return the source type of the device."""
+        """Return the source type."""
         return SOURCE_TYPE_ROUTER
+
+    @property
+    def mac(self) -> str:
+        """Return the MAC address."""
+        return self._mac
+
+    @property
+    def manufacturer(self) -> str:
+        """Return the manufacturer."""
+        return self._manufacturer
 
     @property
     def icon(self) -> str:
         """Return the icon."""
-        return self._device.icon
+        return self._icon
+
+    @property
+    def active(self) -> bool:
+        """Return true if the host sends traffic to the Freebox."""
+        return self._active
+
+    @property
+    def reachable(self) -> bool:
+        """Return true if the host can receive traffic from the Freebox."""
+        return self._reachable
 
     @property
     def device_state_attributes(self) -> Dict[str, any]:
-        """Return the device state attributes."""
-        return self._device.state_attributes
+        """Return the attributes."""
+        return self._attrs
 
     @property
     def device_info(self) -> Dict[str, any]:
         """Return the device information."""
         return {
-            "connections": {(CONNECTION_NETWORK_MAC, self._device.mac)},
+            "connections": {(CONNECTION_NETWORK_MAC, self.mac)},
             "identifiers": {(DOMAIN, self.unique_id)},
             "name": self.name,
-            "manufacturer": self._device.manufacturer,
+            "manufacturer": self.manufacturer,
         }
 
     @property
@@ -100,3 +147,28 @@ class FreeboxTrackerEntity(TrackerEntity):
     async def async_will_remove_from_hass(self):
         """Clean up after entity before removal."""
         self._unsub_dispatcher()
+
+
+def icon_for_freebox_device(device) -> str:
+    """Return a host icon from his type."""
+    switcher = {
+        "freebox_delta": "mdi:television-guide",
+        "freebox_hd": "mdi:television-guide",
+        "freebox_mini": "mdi:television-guide",
+        "freebox_player": "mdi:television-guide",
+        "ip_camera": "mdi:cctv",
+        "ip_phone": "mdi:phone-voip",
+        "laptop": "mdi:laptop",
+        "multimedia_device": "mdi:play-network",
+        "nas": "mdi:nas",
+        "networking_device": "mdi:network",
+        "printer": "mdi:printer",
+        "router": "mdi:router-wireless",
+        "smartphone": "mdi:cellphone",
+        "tablet": "mdi:tablet",
+        "television": "mdi:television",
+        "vg_console": "mdi:gamepad-variant",
+        "workstation": "mdi:desktop-tower-monitor",
+    }
+
+    return switcher.get(device["host_type"], "mdi:help-network")
